Improve `elba test`
  - Allow test binaries to depend on all (dev-)dependencies of the root
    package.

  - OutputLayout's build folder has been further separated into lib/
    and bin/, and test binaries now have lib/ added as an import path.
    This allows test binaries to test non-exported modules.

+++ src/lib/util/mod.rs
@@ -98,7 +98,9 @@ pub fn copy_dir(from: &Path, to: &Path) -> Res<()> {
 }
 
 pub fn clear_dir(dir: &Path) -> Res<()> {
-    fs::remove_dir_all(dir)?;
+    if dir.exists() {
+        fs::remove_dir_all(dir)?;
+    }
     fs::create_dir(dir)?;
     Ok(())
 }
